+++ admin/models/compiler.php
@@ -308,10 +308,10 @@ class ComponentbuilderModelCompiler extends ListModel
 	/**
 	 * Get all components in the system
 	 *
-	 * @return  array
+	 * @return  array|null
 	 * @since   3.2.0
 	 **/
-	public function getComponents(): array
+	public function getComponents(): ?array
 	{
 		// Get a db connection.
 		$db = $this->getDbo();
@@ -334,9 +334,8 @@ class ComponentbuilderModelCompiler extends ListModel
 		$db->setQuery($query);
 
 		// return the result
-		return $db->loadObjectList();
+		return $db->loadObjectList() ?? null;
 	}
-
 
 	/**
 	 * Get all dynamic content