+++ admin/models/search.php
@@ -220,10 +220,10 @@ class ComponentbuilderModelSearch extends ItemModel
 	/**
 	 * Get all components in the system
 	 *
-	 * @return  array
+	 * @return  array|null
 	 * @since   3.2.0
 	 **/
-	public function getComponents(): array
+	public function getComponents(): ?array
 	{
 		// Get a db connection.
 		$db = $this->getDbo();
@@ -246,7 +246,6 @@ class ComponentbuilderModelSearch extends ItemModel
 		$db->setQuery($query);
 
 		// return the result
-		return $db->loadObjectList();
+		return $db->loadObjectList() ?? null;
 	}
-
 }